fix(agent): restore desktop agent IPC module, rename plugin commands
The plugin-commands module was created as core/agent/commands.rs,
clobbering the desktop Tauri IPC surface of the same name and declaring a
duplicate module in non-CLI builds. Restore the IPC surface and move the
plugin commands module to plugin_commands.rs.

Also in this commit:
- hoist the per-plugin artifact scans out of the installed_entries loop
  (one full walk per artifact kind, not per plugin)
- share one invocation-wrapper helper between the skill and command
  producers so resume parsing cannot drift from the written format
