+++ src-tauri/plugins/tauri-plugin-llamacpp/permissions/autogenerated/reference.md
@@ -948,32 +948,6 @@ Denies the prioritize_backends command without any pre-configured scope.
 <tr>
 <td>
 
-`llamacpp:allow-probe-backend-load`
-
-</td>
-<td>
-
-Enables the probe_backend_load command without any pre-configured scope.
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`llamacpp:deny-probe-backend-load`
-
-</td>
-<td>
-
-Denies the probe_backend_load command without any pre-configured scope.
-
-</td>
-</tr>
-
-<tr>
-<td>
-
 `llamacpp:allow-read-gguf-metadata`
 
 </td>
